Fix #2689: reject vector embeddings that underflow to zero in f32
Subnormal-magnitude embedding components (|x| < ~1.2e-38) underflowed to
0.0f32 when serde deserialized the wire `Vec<f32>` fields, and the engine's
finiteness check ran post-cast, so the zero passed and a silently-corrupted
(all-zero) vector was stored with applied=true. The opposite extreme (inf) was
caught, so the two extremes were asymmetric.

Widen the upsert wire embedding fields to f64 (Command::VectorUpsert.vector,
BatchVectorEntry.vector, and the CLI parser) and add engine
VectorEmbedding::from_wire, which narrows f64 to the stored f32 and rejects a
non-zero magnitude that underflows to zero — with the same
invalid_argument.engine.vector_embedding as the overflow case. Stored format is
unchanged (still f32; bit-identical for representable values). Query vectors
keep f32 and are tracked separately (#2710).

Invariants verified: SCALE-003 (HOLDS), ARCH-003 (HOLDS) — ingest validation
only; stored representation and write path unchanged.
Tests added: from_wire_rejects_non_representable_values_and_keeps_genuine_zero, vector_upsert_rejects_a_subnormal_embedding_instead_of_storing_zeros

+++ crates/executor/src/types/vector_input.rs
@@ -136,14 +136,16 @@ impl VectorMetadataFilter {
 #[cfg_attr(feature = "idl-tooling", derive(schemars::JsonSchema))]
 pub struct BatchVectorEntry {
     key: String,
-    vector: Vec<f32>,
+    /// Dense embedding at wire (f64) precision; narrowed to the stored f32 by the
+    /// engine, which rejects values that underflow or overflow f32.
+    vector: Vec<f64>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
     metadata: Option<Value>,
 }
 
 impl BatchVectorEntry {
     /// Creates a vector batch upsert entry.
-    pub fn new(key: impl Into<String>, vector: Vec<f32>, metadata: Option<Value>) -> Self {
+    pub fn new(key: impl Into<String>, vector: Vec<f64>, metadata: Option<Value>) -> Self {
         Self {
             key: key.into(),
             vector,
@@ -157,7 +159,7 @@ impl BatchVectorEntry {
     }
 
     /// Returns the embedding.
-    pub fn vector(&self) -> &[f32] {
+    pub fn vector(&self) -> &[f64] {
         &self.vector
     }
 
@@ -167,7 +169,7 @@ impl BatchVectorEntry {
     }
 
     /// Consumes the entry.
-    pub fn into_parts(self) -> (String, Vec<f32>, Option<Value>) {
+    pub fn into_parts(self) -> (String, Vec<f64>, Option<Value>) {
         (self.key, self.vector, self.metadata)
     }
 }
